+++ controllers/gateway/config_map.go
@@ -18,10 +18,10 @@ func (r *APIRuleReconciler) reconcileConfigMap(ctx context.Context, isCMReconcil
 	err := r.Config.ReadFromConfigMap(ctx, r.Client)
 	if err != nil {
 		if apierrs.IsNotFound(err) {
-			r.Log.Info(fmt.Sprintf(`ConfigMap %s in namespace %s was not found {"controller": "Api"}, will use default config`, configMapName, configMapNamespace))
+			r.Log.Info(fmt.Sprintf(`ConfigMap %s in namespace %s was not found {"controller": "ApiRule"}, will use default config`, configMapName, configMapNamespace))
 			r.Config.ResetToDefault()
 		} else {
-			r.Log.Error(err, fmt.Sprintf(`could not read ConfigMap %s in namespace %s {"controller": "Api"}`, configMapName, configMapNamespace))
+			r.Log.Error(err, fmt.Sprintf(`could not read ConfigMap %s in namespace %s {"controller": "ApiRule"}`, configMapName, configMapNamespace))
 			r.Config.Reset()
 		}
 	}
@@ -30,7 +30,7 @@ func (r *APIRuleReconciler) reconcileConfigMap(ctx context.Context, isCMReconcil
 		r.Log.Info("ConfigMap changed", "config", r.Config)
 		if len(configValidationFailures) > 0 {
 			failuresJson, _ := json.Marshal(configValidationFailures)
-			r.Log.Error(err, fmt.Sprintf(`Config validation failure {"controller": "Api", "failures": %s}`, string(failuresJson)))
+			r.Log.Error(err, fmt.Sprintf(`Config validation failure {"controller": "ApiRule", "failures": %s}`, string(failuresJson)))
 		}
 		r.Log.Info("ConfigMap reconciliation finished")
 		return true
