+++ internal/processing/processors/istio/authorization_policy_processor_test.go
@@ -182,10 +182,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt})
 		svc := GetService(*apiRule.Spec.Service.Name)
 		client := GetFakeClient(svc)
-		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 		Expect(err).To(BeNil())
 		Expect(result).To(HaveLen(1))
@@ -209,10 +209,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt, ruleJwt2})
 		svc := GetService(*apiRule.Spec.Service.Name)
 		client := GetFakeClient(svc)
-		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -283,10 +283,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt})
 		svc := GetService(*apiRule.Spec.Service.Name)
 		client := GetFakeClient(svc)
-		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -347,10 +347,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt})
 		svc := GetService(*apiRule.Spec.Service.Name)
 		client := GetFakeClient(svc)
-		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -376,10 +376,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt}, specServiceNamespace)
 		svc := GetService(ruleServiceName, specServiceNamespace)
 		client := GetFakeClient(svc)
-		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -406,10 +406,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt})
 		svc := GetService(ruleServiceName, ruleServiceNamespace)
 		client := GetFakeClient(svc)
-		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -447,10 +447,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt})
 		svc := GetService(*apiRule.Spec.Service.Name)
 		client := GetFakeClient(svc)
-		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -516,10 +516,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 
 			svc := GetService(overrideServiceName, overrideServiceNamespace)
 			client := GetFakeClient(svc)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -555,10 +555,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			}
 
 			client := GetFakeClient()
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -592,10 +592,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleAllow, ruleJwt})
 			svc := GetService(*apiRule.Spec.Service.Name)
 			client := GetFakeClient(svc)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			results, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			results, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -657,10 +657,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleNoop, ruleJwt})
 			svc := GetService(*apiRule.Spec.Service.Name)
 			client := GetFakeClient(svc)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			results, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			results, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -712,10 +712,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 		svc := GetService(serviceName)
 		client := GetFakeClient(svc)
 
-		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -726,7 +726,6 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 	It("should update AP when path, methods and service name didn't change", func() {
 		// given: Cluster state
 		existingAp := getAuthorizationPolicy("raName", ApiNamespace, "test-service", []string{"GET", "POST"})
-		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 		// given: New resources
 		path := "/"
@@ -739,8 +738,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 		svc := GetService(serviceName)
 		client := GetFakeClient(existingAp, svc)
 
+		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
+
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -758,7 +759,6 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			svc1 := GetService("test-service")
 			svc2 := GetService("jwt-secured-service")
 			ctrlClient := GetFakeClient(beingUpdatedAp, jwtSecuredAp, svc1, svc2)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 			// given: New resources
 			jwtRule := getRuleForApTest(methodsGetPost, "/", "jwt-secured-service")
@@ -782,9 +782,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			rules := []gatewayv1beta1.Rule{rule, jwtRule}
 
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -802,13 +803,13 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 		existingAp := getAuthorizationPolicy("raName", ApiNamespace, "test-service", []string{"GET", "POST"})
 		svc := GetService("test-service")
 		ctrlClient := GetFakeClient(existingAp, svc)
-		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 		// given: New resources
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{})
+		processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+		result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 		// then
 		Expect(err).To(BeNil())
@@ -824,7 +825,6 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			existingAp := getAuthorizationPolicy("raName", ApiNamespace, "test-service", []string{"GET", "POST"})
 			svc := GetService("test-service")
 			ctrlClient := GetFakeClient(existingAp, svc)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 			// given: New resources
 
@@ -833,9 +833,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			rules := []gatewayv1beta1.Rule{existingRule, newRule}
 
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -851,7 +852,6 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			existingAp := getAuthorizationPolicy("raName", ApiNamespace, "test-service", []string{"GET", "POST"})
 			svc := GetService("test-service")
 			ctrlClient := GetFakeClient(existingAp, svc)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 			// given: New resources
 
@@ -860,9 +860,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			rules := []gatewayv1beta1.Rule{existingRule, newRule}
 
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -885,10 +886,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			svc1 := GetService("test-service")
 			svc2 := GetService("new-service")
 			ctrlClient := GetFakeClient(existingAp, svc1, svc2)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -904,7 +905,6 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			existingAp := getAuthorizationPolicy("raName", ApiNamespace, "test-service", []string{"GET", "POST"})
 			svc := GetService("test-service")
 			ctrlClient := GetFakeClient(existingAp, svc)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 			// given: New resources
 			path := "/new-path"
@@ -914,9 +914,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			rules := []gatewayv1beta1.Rule{rule}
 
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -936,7 +937,6 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			toBeUpdateAp := getAuthorizationPolicy("to-be-updated-ap", ApiNamespace, "test-service", []string{"GET"})
 			svc := GetService("test-service")
 			ctrlClient := GetFakeClient(toBeUpdateAp, unchangedAp, svc)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 			// given: New resources
 			unchangedRule := getRuleForApTest(methodsDelete, "/", "test-service")
@@ -944,9 +944,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			rules := []gatewayv1beta1.Rule{updatedRule, unchangedRule}
 
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -967,7 +968,6 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			svc := GetService("test-service")
 			svcNewNS := GetService("test-service", specNewServiceNamespace)
 			ctrlClient := GetFakeClient(oldAP, svc, svcNewNS)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 			// given: New resources
 			movedRule := getRuleForApTest(methodsDelete, "/", "test-service")
@@ -975,9 +975,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 
 			apiRule := GetAPIRuleFor(rules)
 			apiRule.Spec.Service.Namespace = &specNewServiceNamespace
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -993,16 +994,16 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			oldAP := getAuthorizationPolicy("unchanged-ap", ApiNamespace, "test-service", []string{"DELETE"})
 			svc := GetService("test-service", "new-namespace")
 			ctrlClient := GetFakeClient(oldAP, svc)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 			// given: New resources
 			movedRule := getRuleForApTest(methodsDelete, "/", "test-service", "new-namespace")
 			rules := []gatewayv1beta1.Rule{movedRule}
 
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -1022,7 +1023,6 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			svc1 := GetService("first-service")
 			svc2 := GetService("second-service")
 			ctrlClient := GetFakeClient(toBeUpdateAp, unchangedAp, svc1, svc2)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 			// given: New resources
 			unchangedRule := getRuleForApTest(methodsGet, "/", "first-service")
@@ -1030,9 +1030,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			rules := []gatewayv1beta1.Rule{updatedRule, unchangedRule}
 
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -1070,7 +1071,6 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 
 			svc := GetService(serviceName)
 			ctrlClient := GetFakeClient(ap1, ap2, svc)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 			// given: ApiRule with updated audiences in jwt authorizations
 			authorization1 := `{"audiences": ["audience1", "audience3"]}`
@@ -1095,9 +1095,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			rules := []gatewayv1beta1.Rule{rule}
 
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -1135,7 +1136,6 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 
 			svc := GetService(serviceName)
 			ctrlClient := GetFakeClient(ap1, ap2, svc)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 			// given: ApiRule with updated audiences in jwt authorizations
 			authorization1 := `{"audiences": ["audience1", "audience2"]}`
@@ -1161,9 +1161,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			rules := []gatewayv1beta1.Rule{rule}
 
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -1202,7 +1203,6 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 
 			svc := GetService(serviceName)
 			ctrlClient := GetFakeClient(ap1, ap2, svc)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 			// given: ApiRule with updated audiences in jwt authorizations
 			authorization1 := `{"audiences": ["audience1", "audience2"]}`
@@ -1226,9 +1226,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			rules := []gatewayv1beta1.Rule{rule}
 
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -1278,7 +1279,6 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 
 			svc := GetService(serviceName)
 			ctrlClient := GetFakeClient(ap1, ap2, ap3, svc)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 			// given: ApiRule with updated audiences in jwt authorizations
 			authorization2 := `{"audiences": ["audience3"]}`
@@ -1303,9 +1303,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			rules := []gatewayv1beta1.Rule{rule}
 
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -1335,10 +1336,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			svc.Spec.Selector["custom"] = serviceName
 			client := GetFakeClient(svc)
 
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -1366,10 +1367,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			svc.Spec.Selector["custom"] = serviceName
 			client := GetFakeClient(svc)
 
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -1396,10 +1397,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			svc.Spec.Selector["second-custom"] = "blah"
 			client := GetFakeClient(svc)
 
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -1435,7 +1436,6 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 
 			svc := GetService(serviceName)
 			ctrlClient := GetFakeClient(ap, svc)
-			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger)
 
 			// given: ApiRule with updated audiences in jwt authorizations
 			authorization := `{"audiences": ["audience1"]}`
@@ -1459,9 +1459,10 @@ var _ = Describe("JwtAuthorization Policy Processor", func() {
 			rules := []gatewayv1beta1.Rule{rule}
 
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1AuthorizationPolicyProcessor(GetTestConfig(), &testLogger, apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
