+++ internal/processing/processors/istio/request_authentication_processor_test.go
@@ -94,10 +94,10 @@ var _ = Describe("Request Authentication Processor", func() {
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt, ruleJwt2})
 		svc := GetService(ServiceName)
 		client := GetFakeClient(svc)
-		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
+		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -122,10 +122,10 @@ var _ = Describe("Request Authentication Processor", func() {
 		client := GetFakeClient(svc)
 		ruleJwt := GetRuleFor(HeadersApiPath, ApiMethods, []*gatewayv1beta1.Mutator{}, []*gatewayv1beta1.Authenticator{jwt})
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt})
-		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
+		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -151,10 +151,10 @@ var _ = Describe("Request Authentication Processor", func() {
 		client := GetFakeClient(svc)
 		ruleJwt := GetRuleWithServiceFor(HeadersApiPath, ApiMethods, []*gatewayv1beta1.Mutator{}, []*gatewayv1beta1.Authenticator{jwt}, service)
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt}, specServiceNamespace)
-		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
+		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -181,10 +181,10 @@ var _ = Describe("Request Authentication Processor", func() {
 		client := GetFakeClient(svc)
 		ruleJwt := GetRuleWithServiceFor(HeadersApiPath, ApiMethods, []*gatewayv1beta1.Mutator{}, []*gatewayv1beta1.Authenticator{jwt}, service)
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt})
-		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
+		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -220,10 +220,10 @@ var _ = Describe("Request Authentication Processor", func() {
 		}
 		ruleJwt := GetRuleWithServiceFor(HeadersApiPath, ApiMethods, []*gatewayv1beta1.Mutator{}, []*gatewayv1beta1.Authenticator{jwt}, service)
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt})
-		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
+		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -270,10 +270,10 @@ var _ = Describe("Request Authentication Processor", func() {
 		}
 
 		client := GetFakeClient()
-		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
+		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -291,10 +291,10 @@ var _ = Describe("Request Authentication Processor", func() {
 		svc := GetService("test-service")
 		client := GetFakeClient(svc)
 		apiRule := GetAPIRuleFor(rules)
-		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
+		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+		result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 		// then
 		Expect(err).To(BeNil())
@@ -307,13 +307,13 @@ var _ = Describe("Request Authentication Processor", func() {
 		existingRa := getRequestAuthentication("raName", "test-service", JwksUri, JwtIssuer)
 
 		ctrlClient := GetFakeClient(&existingRa)
-		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
 
 		// given: New resources
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{})
+		processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 		// when
-		result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+		result, err := processor.EvaluateReconciliation(context.Background(), ctrlClient)
 
 		// then
 		Expect(err).To(BeNil())
@@ -328,15 +328,15 @@ var _ = Describe("Request Authentication Processor", func() {
 			existingRa := getRequestAuthentication("raName", "test-service", JwksUri, JwtIssuer)
 			svc := GetService("test-service")
 			ctrlClient := GetFakeClient(&existingRa, svc)
-			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
 
 			// given: New resources
 			jwtRule := GetJwtRuleWithService(JwtIssuer, JwksUri, "test-service")
 			rules := []gatewayv1beta1.Rule{jwtRule}
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -350,15 +350,15 @@ var _ = Describe("Request Authentication Processor", func() {
 			svcOld := GetService("old-service")
 			svcUpdated := GetService("updated-service")
 			ctrlClient := GetFakeClient(&existingRa, svcOld, svcUpdated)
-			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
 
 			// given: New resources
 			jwtRule := GetJwtRuleWithService(JwtIssuer, JwksUri, "updated-service")
 			rules := []gatewayv1beta1.Rule{jwtRule}
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -375,16 +375,16 @@ var _ = Describe("Request Authentication Processor", func() {
 			svcExisting := GetService("existing-service")
 			svcNew := GetService("new-service")
 			ctrlClient := GetFakeClient(&existingRa, svcExisting, svcNew)
-			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
 
 			// given: New resources
 			existingJwtRule := GetJwtRuleWithService(JwtIssuer, JwksUri, "existing-service")
 			newJwtRule := GetJwtRuleWithService("https://new.issuer.com/", JwksUri, "new-service")
 			rules := []gatewayv1beta1.Rule{existingJwtRule, newJwtRule}
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -400,15 +400,15 @@ var _ = Describe("Request Authentication Processor", func() {
 			existingRa := getRequestAuthentication("raName", "test-service", JwksUri, JwtIssuer)
 			svc := GetService("test-service")
 			ctrlClient := GetFakeClient(&existingRa, svc)
-			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
 
 			// given: New resources
 			jwtRule := GetJwtRuleWithService(JwtIssuer, JwksUri2, "test-service")
 			rules := []gatewayv1beta1.Rule{jwtRule}
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -430,16 +430,16 @@ var _ = Describe("Request Authentication Processor", func() {
 			svcFirst := GetService("first-service")
 			svcSecond := GetService("second-service")
 			ctrlClient := GetFakeClient(&firstServiceRa, &secondServiceRa, svcFirst, svcSecond)
-			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
 
 			// given: New resources
 			firstJwtRule := GetJwtRuleWithService("https://new.issuer.com/", JwksUri, "first-service")
 			secondJwtRule := GetJwtRuleWithService(JwtIssuer, JwksUri, "second-service")
 			rules := []gatewayv1beta1.Rule{firstJwtRule, secondJwtRule}
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -458,15 +458,15 @@ var _ = Describe("Request Authentication Processor", func() {
 			svcFirst := GetService("first-service")
 			svcSecond := GetService("second-service")
 			ctrlClient := GetFakeClient(&firstServiceRa, &secondServiceRa, svcFirst, svcSecond)
-			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
 
 			// given: New resources
 			secondJwtRule := GetJwtRuleWithService(JwtIssuer, JwksUri, "second-service")
 			rules := []gatewayv1beta1.Rule{secondJwtRule}
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -485,7 +485,6 @@ var _ = Describe("Request Authentication Processor", func() {
 			svcSecond := GetService("second-service")
 			svcNew := GetService("new-service")
 			ctrlClient := GetFakeClient(&firstServiceRa, &secondServiceRa, svcFirst, svcSecond, svcNew)
-			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
 
 			// given: New resources
 			firstJwtRule := GetJwtRuleWithService(JwtIssuer, JwksUri, "first-service")
@@ -493,9 +492,10 @@ var _ = Describe("Request Authentication Processor", func() {
 			newJwtRule := GetJwtRuleWithService(JwtIssuer, JwksUri, "new-service")
 			rules := []gatewayv1beta1.Rule{firstJwtRule, secondJwtRule, newJwtRule}
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -513,7 +513,6 @@ var _ = Describe("Request Authentication Processor", func() {
 			svcOld := GetService("old-service")
 			svcNewNS := GetService("old-service", "new-namespace")
 			ctrlClient := GetFakeClient(&oldRa, svcOld, svcNewNS)
-			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
 
 			// given: New resources
 			jwtRule := GetJwtRuleWithService(JwtIssuer, JwksUri, "old-service", "new-namespace")
@@ -521,9 +520,10 @@ var _ = Describe("Request Authentication Processor", func() {
 			apiRule := GetAPIRuleFor(rules)
 			specServiceNamespace := "new-namespace"
 			apiRule.Spec.Service.Namespace = &specServiceNamespace
+			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -540,15 +540,15 @@ var _ = Describe("Request Authentication Processor", func() {
 			svcOld := GetService("old-service")
 			svcNewNS := GetService("old-service", "new-namespace")
 			ctrlClient := GetFakeClient(&oldRa, svcOld, svcNewNS)
-			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
 
 			// given: New resources
 			jwtRule := GetJwtRuleWithService(JwtIssuer, JwksUri, "old-service", "new-namespace")
 			rules := []gatewayv1beta1.Rule{jwtRule}
 			apiRule := GetAPIRuleFor(rules)
+			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), ctrlClient, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), ctrlClient)
 
 			// then
 			Expect(err).To(BeNil())
@@ -574,10 +574,10 @@ var _ = Describe("Request Authentication Processor", func() {
 			svc.Spec.Selector["custom"] = serviceName
 			client := GetFakeClient(svc)
 
-			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
+			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -605,10 +605,10 @@ var _ = Describe("Request Authentication Processor", func() {
 			svc.Spec.Selector["custom"] = serviceName
 			client := GetFakeClient(svc)
 
-			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
+			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -635,10 +635,10 @@ var _ = Describe("Request Authentication Processor", func() {
 			svc.Spec.Selector["second-custom"] = "blah"
 			client := GetFakeClient(svc)
 
-			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig())
+			processor := istio.Newv1beta1RequestAuthenticationProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
