+++ internal/processing/processors/ory/virtual_service_processor_test.go
@@ -41,10 +41,10 @@ var _ = Describe("Virtual Service Processor", func() {
 
 				apiRule := GetAPIRuleFor(rules)
 				client := GetFakeClient()
-				processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+				processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 				// when
-				result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+				result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 				// then
 				Expect(err).To(BeNil())
@@ -100,10 +100,10 @@ var _ = Describe("Virtual Service Processor", func() {
 					Port:      &overrideServicePort,
 				}
 				client := GetFakeClient()
-				processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+				processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 				// when
-				result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+				result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 				// then
 				Expect(err).To(BeNil())
@@ -140,10 +140,10 @@ var _ = Describe("Virtual Service Processor", func() {
 
 				apiRule := GetAPIRuleFor(rules)
 				client := GetFakeClient()
-				processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+				processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 				// when
-				result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+				result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 				// then
 				Expect(err).To(BeNil())
@@ -171,10 +171,10 @@ var _ = Describe("Virtual Service Processor", func() {
 				apiRule := GetAPIRuleFor(rules)
 				apiRule.Spec.Host = &ServiceHostWithNoDomain
 				client := GetFakeClient()
-				processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+				processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 				// when
-				result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+				result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 				// then
 				Expect(err).To(BeNil())
@@ -215,10 +215,10 @@ var _ = Describe("Virtual Service Processor", func() {
 
 			apiRule := GetAPIRuleFor(rules)
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -271,10 +271,10 @@ var _ = Describe("Virtual Service Processor", func() {
 				Expect(err).NotTo(HaveOccurred())
 
 				client := fake.NewClientBuilder().WithScheme(scheme).WithObjects(&rule, &vs).Build()
-				processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+				processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 				// when
-				result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+				result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 				// then
 				Expect(err).To(BeNil())
@@ -352,10 +352,10 @@ var _ = Describe("Virtual Service Processor", func() {
 
 			apiRule := GetAPIRuleFor(rules)
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -440,10 +440,10 @@ var _ = Describe("Virtual Service Processor", func() {
 
 			apiRule := GetAPIRuleFor(rules)
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -519,10 +519,10 @@ var _ = Describe("Virtual Service Processor", func() {
 
 			apiRule := GetAPIRuleFor(rules)
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -596,10 +596,10 @@ var _ = Describe("Virtual Service Processor", func() {
 
 			apiRule := GetAPIRuleFor(rules)
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -652,10 +652,10 @@ var _ = Describe("Virtual Service Processor", func() {
 
 			apiRule := GetAPIRuleFor(rules)
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -683,10 +683,10 @@ var _ = Describe("Virtual Service Processor", func() {
 			apiRule := GetAPIRuleFor(rules)
 			apiRule.Spec.Timeout = &timeout10s
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -715,10 +715,10 @@ var _ = Describe("Virtual Service Processor", func() {
 			apiRule := GetAPIRuleFor(rules)
 			apiRule.Spec.Timeout = &timeout10s
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -747,10 +747,10 @@ var _ = Describe("Virtual Service Processor", func() {
 			apiRule := GetAPIRuleFor(rules)
 			apiRule.Spec.Timeout = &timeout10s
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -779,10 +779,10 @@ var _ = Describe("Virtual Service Processor", func() {
 
 			apiRule := GetAPIRuleFor(rules)
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -813,10 +813,10 @@ var _ = Describe("Virtual Service Processor", func() {
 			apiRule := GetAPIRuleFor(rules)
 			apiRule.Spec.Host = &ServiceHostWithNoDomain
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -859,10 +859,10 @@ var _ = Describe("Virtual Service Processor", func() {
 			apiRule.Spec.CorsPolicy = ptr.To(corsPolicy)
 
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -909,10 +909,10 @@ var _ = Describe("Virtual Service Processor", func() {
 			apiRule.Spec.CorsPolicy = ptr.To(corsPolicy)
 
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -963,10 +963,10 @@ var _ = Describe("Virtual Service Processor", func() {
 			apiRule.Spec.CorsPolicy = ptr.To(corsPolicy)
 
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -1016,10 +1016,10 @@ var _ = Describe("Virtual Service Processor", func() {
 
 			apiRule := GetAPIRuleFor(rules)
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
@@ -1057,10 +1057,10 @@ var _ = Describe("Virtual Service Processor", func() {
 
 			apiRule := GetAPIRuleFor(rules)
 			client := GetFakeClient()
-			processor := ory.NewVirtualServiceProcessor(GetTestConfig())
+			processor := ory.NewVirtualServiceProcessor(GetTestConfig(), apiRule)
 
 			// when
-			result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
+			result, err := processor.EvaluateReconciliation(context.Background(), client)
 
 			// then
 			Expect(err).To(BeNil())
