+++ internal/processing/processors/request_authentication_processor.go
@@ -14,7 +14,7 @@ const RequestAuthenticationAppSelectorLabel = "app"
 
 // RequestAuthenticationProcessor is the generic processor that handles the Istio Request Authentications in the reconciliation of API Rule.
 type RequestAuthenticationProcessor struct {
-	Api     *gatewayv1beta1.APIRule
+	ApiRule *gatewayv1beta1.APIRule
 	Creator RequestAuthenticationCreator
 }
 
@@ -25,11 +25,11 @@ type RequestAuthenticationCreator interface {
 }
 
 func (r RequestAuthenticationProcessor) EvaluateReconciliation(ctx context.Context, client ctrlclient.Client) ([]*processing.ObjectChange, error) {
-	desired, err := r.getDesiredState(ctx, client, r.Api)
+	desired, err := r.getDesiredState(ctx, client, r.ApiRule)
 	if err != nil {
 		return make([]*processing.ObjectChange, 0), err
 	}
-	actual, err := r.getActualState(ctx, client, r.Api)
+	actual, err := r.getActualState(ctx, client, r.ApiRule)
 	if err != nil {
 		return make([]*processing.ObjectChange, 0), err
 	}
