+++ internal/processing/processors/v2alpha1/status.go
@@ -1,13 +1,28 @@
 package v2alpha1
 
 import (
-	"github.com/kyma-project/api-gateway/apis/gateway/v2alpha1"
-	v1alpha2Status "github.com/kyma-project/api-gateway/internal/processing/status"
+	gatewayv1beta1 "github.com/kyma-project/api-gateway/apis/gateway/v1beta1"
+	"github.com/kyma-project/api-gateway/internal/processing/status"
+	v1beta1Status "github.com/kyma-project/api-gateway/internal/processing/status"
 )
 
-func Base(state string) v1alpha2Status.ReconciliationV2alpha1Status {
-	return v1alpha2Status.ReconciliationV2alpha1Status{
-		State:       v2alpha1.State(state),
-		Description: "",
+func (r Reconciliation) GetStatusBase(statusCode string) status.ReconciliationStatus {
+	return StatusBase(statusCode)
+}
+
+func StatusBase(statusCode string) status.ReconciliationStatus {
+	return v1beta1Status.ReconciliationV1beta1Status{
+		ApiRuleStatus: &gatewayv1beta1.APIRuleResourceStatus{
+			Code: gatewayv1beta1.StatusCode(statusCode),
+		},
+		VirtualServiceStatus: &gatewayv1beta1.APIRuleResourceStatus{
+			Code: gatewayv1beta1.StatusCode(statusCode),
+		},
+		AuthorizationPolicyStatus: &gatewayv1beta1.APIRuleResourceStatus{
+			Code: gatewayv1beta1.StatusCode(statusCode),
+		},
+		RequestAuthenticationStatus: &gatewayv1beta1.APIRuleResourceStatus{
+			Code: gatewayv1beta1.StatusCode(statusCode),
+		},
 	}
 }
