+++ internal/processing/processors/v2alpha1/status_test.go
@@ -1,16 +1,17 @@
-package istio
+package v2alpha1_test
 
 import (
 	gatewayv1beta1 "github.com/kyma-project/api-gateway/apis/gateway/v1beta1"
+	"github.com/kyma-project/api-gateway/internal/processing/processors/v2alpha1"
 	status "github.com/kyma-project/api-gateway/internal/processing/status"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
 
-var _ = Describe("IstioStatusBase", func() {
+var _ = Describe("StatusBase", func() {
 	It("should create status base with AccessRule set to nil", func() {
 		// when
-		status := StatusBase(string(gatewayv1beta1.StatusSkipped)).(status.ReconciliationV1beta1Status)
+		status := v2alpha1.StatusBase(string(gatewayv1beta1.StatusSkipped)).(status.ReconciliationV1beta1Status)
 
 		Expect(status.ApiRuleStatus.Code).To(Equal(gatewayv1beta1.StatusSkipped))
 		Expect(status.VirtualServiceStatus.Code).To(Equal(gatewayv1beta1.StatusSkipped))
