+++ internal/validation/v1beta1/ory/ory.go
@@ -8,7 +8,7 @@ import (
 
 func NewAPIRuleValidator(api *gatewayv1beta1.APIRule, defaultDomainName string) validation.ApiRuleValidator {
 	return &v1beta1.APIRuleValidator{
-		Api: api,
+		ApiRule: api,
 
 		HandlerValidator:          &HandlerValidator{},
 		AccessStrategiesValidator: &AccessStrategyValidator{},
